+++ src/lib.rs
@@ -71,15 +71,13 @@ pub struct ImGui {
 #[macro_export]
 macro_rules! im_str {
     ($e:tt) => ({
-        debug_assert!(!($e).as_bytes().contains(&0));
         let value = concat!($e, "\0");
-        unsafe { ::imgui::ImStr::from_bytes_unchecked(value.as_bytes()) }
+        unsafe { ::imgui::ImStr::from_utf8_with_nul_unchecked(value.as_bytes()) }
     });
     ($e:tt, $($arg:tt)*) => ({
         let mut bytes: Vec<u8> = format!($e, $($arg)*).into();
-        debug_assert!(!bytes.contains(&0));
         bytes.push(b'\0');
-        unsafe { &::imgui::ImString::from_vec_unchecked(bytes) }
+        unsafe { &::imgui::ImString::from_utf8_unchecked(bytes) }
     })
 }